HEEP-352 Set address to use 16 bit numbers instead of 8 bit

+++ Firmware/PIC18f26k40MCCTest.X/W5500.h
@@ -24,13 +24,16 @@ void ReadDefaultGateway(uint8_t* buf);
 void SetSocketRXSize(uint8_t socket, uint8_t size);
 void SetSocketTXSize(uint8_t socket, uint8_t size);
 
-void SetSingleByteW5500(uint8_t addr, uint8_t value);
-uint8_t ReadSingleByteW5500(uint8_t addr);
+void SetSingleByteW5500(uint16_t addr, uint8_t value);
+uint8_t ReadSingleByteW5500(uint16_t addr);
 
 void SetMR(uint8_t newVal);
 uint8_t ReadMR();
 
 uint8_t ReadPHYCFGR();
+
+
+
 
 void WriteDestinationIP(uint8_t socket, uint8_t* buf);
 void WriteDestinationPort(uint8_t socket, uint8_t* buf);
